move src files to /src/service-worker
